+++ Locker/Plugin.cs
@@ -15,7 +15,7 @@ namespace Locker
     {
         public const string ModGUID = "com.zealsprince.locker";
         public const string ModName = "Locker";
-        public const string ModVersion = "0.11.2";
+        public const string ModVersion = "0.12.0";
 
         // These need to be lowercase because we're passing through the protected properties.
         public static ManualLogSource logger;